Added optional initialization functions for service modules.
Flaw used to check available information on the system before loading
the battery module (and the alsa module should have proposed a similar
function for a long time). Now, this kind of check can be offered
directly from the service module using the init function.

The init function, if it exists, is called when the service module is
loaded. If the module cannot be used on the system, it should return
nil. Otherwise, it should proceed to the registration of the gadgets
and return the module itself.

The init.lua code module also gets a better service modules
tracking. For now, is still only reports the modules not loaded, if
any.